Addition of pRESTige framework
This is a MySQL to RESTful API generator which is so magical that you will be amazed. Automatic API generation along with documentation in Swagger Open API specifications. Bundled Code and DB editors and a lot of goodies to kickstart a new project.

There are also a lot of FrontEnd examples showcasing the usage of backend in an efficient manner.